requirements.txt
Django==4.2
djangorestframework
djongo
pymongo[srv]==4.5.0
python-dotenv==1.0.0
sqlparse
